requirements.txt
python-dotenv
aiogram==2.25.1
gspread
oauth2client
